+++ src/components/label/label.styles.ts
@@ -1,0 +1,91 @@
+import { css } from 'lit';
+import componentStyles from '../../styles/component.styles';
+
+export default css`
+  ${componentStyles}
+
+  .label--primary {
+    color: var(--cps-foreground-primary);
+  }
+
+  .label--secondary {
+    color: var(--cps-foreground-secondary);
+  }
+
+  .label--tertiary {
+    color: var(--cps-foreground-tertiary);
+  }
+
+  .label--disabled {
+    color: var(--cps-foreground-disabled);
+  }
+
+  .label--brand-primary {
+    color: var(--cps-foreground-brand-primary);
+  }
+
+  .label--brand-secondary {
+    color: var(--cps-foreground-brand-secondary);
+  }
+
+  .label--brand-tertiary {
+    color: var(--cps-foreground-brand-tertiary);
+  }
+
+  .label--inverted-primary {
+    color: var(--cps-foreground-inverted-primary);
+  }
+
+  .label--inverted-secondary {
+    color: var(--cps-foreground-inverted-secondary);
+  }
+
+  .label--inverted-disabled {
+    color: var(--cps-foreground-inverted-disabled);
+  }
+
+  /* Size modifiers */
+  .label--stamp {
+    font: var(--cps-text-stamp);
+  }
+
+  .label--caption {
+    font: var(--cps-text-caption);
+  }
+
+  .label--label {
+    font: var(--cps-text-label);
+  }
+
+  .label--body {
+    font: var(--cps-text-body);
+  }
+
+  .label--body-em {
+    font: var(--cps-text-body-em);
+  }
+
+  .label--body-strong {
+    font: var(--cps-text-body-strong);
+  }
+
+  .label--body-large {
+    font: var(--cps-text-body-large);
+  }
+
+  .label--subtitle {
+    font: var(--cps-text-subtitle);
+  }
+
+  .label--title {
+    font: var(--cps-text-title);
+  }
+
+  .label--heading {
+    font: var(--cps-text-heading);
+  }
+
+  .label--display {
+    font: var(--cps-text-display);
+  }
+`;