+++ CHANGELOG.md
@@ -6,6 +6,10 @@ O versionamento deste projeto é aderente aos princípios de [Semantic Versionin
 
 ## Unreleased
 
+### Added
+
+- Criação do componente `<cps-label>`.
+
 ## 0.5.0 - 2023-06-15
 
 ### Added